requirements.txt
numpy==2.2.3
scipy==1.15.1
pandas==2.2.3
mne[hdf5]==1.9.0
openpyxl==3.1.5
scikit-learn==1.6.1
pingouin==0.5.5
xgboost==2.1.4
dtw-python==1.5.3
pillow==11.1.0
